Move the Retry function to its own "internal/tunnel" package
and fix British->American English typo!
